# frontend/src/openapi/openapi.json
{"openapi":"3.0.1","info":{"title":"OpenApi specification","description":"OpenApi documentation for shortening urls","contact":{"name":"API Support"},"version":"1.0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"tags":[{"name":"URL Shortening APIs","description":"APIs for shortening, managing, and resolving URLs. Includes endpoints for creating short URLs, retrieving original URLs, and analyzing URL usage statistics."}],"paths":{"/profile":{"get":{"tags":["profile-controller"],"operationId":"listAllFormsOfMetadata","responses":{"200":{"description":"OK","content":{"application/hal+json":{"schema":{"$ref":"#/components/schemas/RepresentationModelObject"}}}}}}},"/profile/urlShorteneds":{"get":{"tags":["profile-controller"],"operationId":"descriptor","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}},"application/alps+json":{"schema":{"type":"string"}},"application/schema+json":{"schema":{"$ref":"#/components/schemas/JsonSchema"}}}}}}},"/urlShorteneds":{"get":{"tags":["url-shortened-entity-controller"],"description":"get-urlshortened","operationId":"getCollectionResource-urlshortened-get","parameters":[{"name":"page","in":"query","description":"Zero-based page index (0..N)","required":false,"schema":{"minimum":0,"type":"integer","default":0}},{"name":"size","in":"query","description":"The size of the page to be returned","required":false,"schema":{"minimum":1,"type":"integer","default":20}},{"name":"sort","in":"query","description":"Sorting criteria in the format: property,(asc|desc). Default sort order is ascending. Multiple sort criteria are supported.","required":false,"schema":{"type":"array","items":{"type":"string"}}}],"responses":{"200":{"description":"OK","content":{"application/hal+json":{"schema":{"$ref":"#/components/schemas/PagedModelEntityModelUrlShortened"}},"application/x-spring-data-compact+json":{"schema":{"$ref":"#/components/schemas/PagedModelEntityModelUrlShortened"}},"text/uri-list":{"schema":{"type":"string"}}}}}},"post":{"tags":["url-shortened-entity-controller"],"description":"create-urlshortened","operationId":"postCollectionResource-urlshortened-post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UrlShortenedRequestBody"}}},"required":true},"responses":{"201":{"description":"Created","content":{"application/hal+json":{"schema":{"$ref":"#/components/schemas/EntityModelUrlShortened"}}}}}}},"/urlShorteneds/search/findByShortenedUrl":{"get":{"tags":["url-shortened-search-controller"],"operationId":"executeSearch-urlshortened-get","parameters":[{"name":"shortenedUrl","in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"application/hal+json":{"schema":{"$ref":"#/components/schemas/EntityModelUrlShortened"}}}},"404":{"description":"Not Found"}}}},"/urlShorteneds/{id}":{"get":{"tags":["url-shortened-entity-controller"],"description":"get-urlshortened","operationId":"getItemResource-urlshortened-get","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"application/hal+json":{"schema":{"$ref":"#/components/schemas/EntityModelUrlShortened"}}}},"404":{"description":"Not Found"}}},"put":{"tags":["url-shortened-entity-controller"],"description":"update-urlshortened","operationId":"putItemResource-urlshortened-put","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UrlShortenedRequestBody"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/hal+json":{"schema":{"$ref":"#/components/schemas/EntityModelUrlShortened"}}}},"201":{"description":"Created","content":{"application/hal+json":{"schema":{"$ref":"#/components/schemas/EntityModelUrlShortened"}}}},"204":{"description":"No Content"}}},"delete":{"tags":["url-shortened-entity-controller"],"description":"delete-urlshortened","operationId":"deleteItemResource-urlshortened-delete","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"204":{"description":"No Content"},"404":{"description":"Not Found"}}},"patch":{"tags":["url-shortened-entity-controller"],"description":"patch-urlshortened","operationId":"patchItemResource-urlshortened-patch","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UrlShortenedRequestBody"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/hal+json":{"schema":{"$ref":"#/components/schemas/EntityModelUrlShortened"}}}},"204":{"description":"No Content"}}}},"/shorten/update/{id}":{"put":{"tags":["URL Shortening APIs"],"summary":"update an url by it's id","operationId":"updateUrl","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RequestUrl"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/hal+json":{"schema":{"$ref":"#/components/schemas/UrlShortened"}}}}}}},"/shorten/customUrl":{"post":{"tags":["URL Shortening APIs"],"summary":"create a custom short url","operationId":"createCustomUrl","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CustomUrlRequest"}}},"required":true},"responses":{"200":{"description":"OK"}}}},"/shorten/create":{"post":{"tags":["URL Shortening APIs"],"summary":"create a new url","operationId":"createUrl","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RequestUrl"}}},"required":true},"responses":{"200":{"description":"OK"}}}},"/shorten/copy/{id}":{"post":{"tags":["URL Shortening APIs"],"summary":"copy url to clipboard","operationId":"copyUrl","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK"}}}},"/shorten/":{"get":{"tags":["URL Shortening APIs"],"summary":"get all urls","operationId":"getAllUrls","responses":{"200":{"description":"OK","content":{"application/hal+json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/UrlShortened"}}}}}}}},"/shorten/GetOriginal":{"get":{"tags":["URL Shortening APIs"],"summary":"get an original url","operationId":"getOriginalUrl","parameters":[{"name":"shortUrl","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"application/hal+json":{"schema":{"$ref":"#/components/schemas/UrlShortened"}}}}}}},"/shorten/Accessed/{id}":{"get":{"tags":["URL Shortening APIs"],"summary":"get number of accessed url","operationId":"accessedUrl","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"application/hal+json":{"schema":{"type":"integer","format":"int32"}}}}}}},"/shorten/delete/{id}":{"delete":{"tags":["URL Shortening APIs"],"summary":"delete an url by it's id","operationId":"deleteUrl","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK"}}}}},"components":{"schemas":{"AbstractJsonSchemaPropertyObject":{"type":"object","properties":{"title":{"type":"string"},"readOnly":{"type":"boolean"}}},"Item":{"type":"object","properties":{"type":{"type":"string"},"properties":{"type":"object","additionalProperties":{"$ref":"#/components/schemas/AbstractJsonSchemaPropertyObject"}},"requiredProperties":{"type":"array","items":{"type":"string"}}}},"JsonSchema":{"type":"object","properties":{"title":{"type":"string"},"description":{"type":"string"},"properties":{"type":"object","additionalProperties":{"$ref":"#/components/schemas/AbstractJsonSchemaPropertyObject"}},"requiredProperties":{"type":"array","items":{"type":"string"}},"definitions":{"type":"object","additionalProperties":{"$ref":"#/components/schemas/Item"}},"type":{"type":"string"},"$schema":{"type":"string"}}},"Links":{"type":"object","additionalProperties":{"$ref":"#/components/schemas/Link"}},"RepresentationModelObject":{"type":"object","properties":{"_links":{"$ref":"#/components/schemas/Links"}}},"EntityModelUrlShortened":{"type":"object","properties":{"originalUrl":{"type":"string"},"shortenedUrl":{"type":"string"},"createdAt":{"type":"string","format":"date-time"},"updatedAt":{"type":"string","format":"date-time"},"accessCount":{"type":"integer","format":"int32"},"type":{"type":"string"},"_links":{"$ref":"#/components/schemas/Links"}}},"PageMetadata":{"type":"object","properties":{"size":{"type":"integer","format":"int64"},"totalElements":{"type":"integer","format":"int64"},"totalPages":{"type":"integer","format":"int64"},"number":{"type":"integer","format":"int64"}}},"PagedModelEntityModelUrlShortened":{"type":"object","properties":{"_embedded":{"type":"object","properties":{"urlShorteneds":{"type":"array","items":{"$ref":"#/components/schemas/EntityModelUrlShortened"}}}},"_links":{"$ref":"#/components/schemas/Links"},"page":{"$ref":"#/components/schemas/PageMetadata"}}},"UrlShortened":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"originalUrl":{"type":"string"},"shortenedUrl":{"type":"string"},"createdAt":{"type":"string","format":"date-time"},"updatedAt":{"type":"string","format":"date-time"},"accessCount":{"type":"integer","format":"int32"},"type":{"type":"string"}}},"UrlShortenedRequestBody":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"originalUrl":{"type":"string"},"shortenedUrl":{"type":"string"},"createdAt":{"type":"string","format":"date-time"},"updatedAt":{"type":"string","format":"date-time"},"accessCount":{"type":"integer","format":"int32"},"type":{"type":"string"}}},"RequestUrl":{"required":["originalUrl"],"type":"object","properties":{"originalUrl":{"pattern":"^(https?|ftp)://[^\\s/$.?#].[^\\s]*$","type":"string"}}},"CustomUrlRequest":{"required":["originalUrl","shortUrl"],"type":"object","properties":{"shortUrl":{"maxLength":2147483647,"minLength":4,"type":"string"},"originalUrl":{"pattern":"^(https?|ftp)://[^\\s/$.?#].[^\\s]*$","type":"string"}}},"Link":{"type":"object","properties":{"href":{"type":"string"},"hreflang":{"type":"string"},"title":{"type":"string"},"type":{"type":"string"},"deprecation":{"type":"string"},"profile":{"type":"string"},"name":{"type":"string"},"templated":{"type":"boolean"}}}}}}